+++ Src/modules/imu/imu.hpp
@@ -73,7 +73,9 @@ private:
         return raw_accel * 9.80665f / 16384.0f;
     }
     constexpr int16_t raw_temp_convert_to_celsius (int16_t raw_temp){
-        temp = raw_temp / 333.87f + 21.0f;
+        // IMPORTANT: this temperature is "die" (iternal) temperature of the MPU9250. It DOES NOT refer to ambient temperature
+        // 21 is from Room Temp Offset 
+        temp = (float)raw_temp / 333.87f + 21.0;
         // Simple and low cost rounding
         return (static_cast<int16_t>(temp + (temp >= 0 ? 0.5f : -0.5f)));
     }
